+++ extension/css/style.css
@@ -1,0 +1,3 @@
+bdi.tempoSalario{
+	border-bottom: 1px dashed rgba(128,128,128,0.1);
+}
