+++ extension/index.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html >
+<head>
+  <meta charset="UTF-8">
+  <script src="js/index.js"></script> 
+</head>
+<body>
+  <div class="wrapper" style="width:150px;">
+    <input type="radio" id="on" name="type" value="ligado" checked>
+    <label for="on">Ligado</label><br>
+    
+    <input type="radio" id="off" name="type" value="desligado">
+    <label for="off">Desligado</label><br>
+    
+    <p class="texto">Clique com o direito e em opções para alterar valores de horas e de salário</p>
+  </div>
+   
+</body>
+</html>
