--- mais-valia.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-  <html>
-    <head>
-      <meta charset="UTF-8">
-      <link rel="stylesheet" type="text/css" href="/css/options.css">
-      <title>
-        O que compra a sua hora - Configurações
-      </title>
-    </head>
-    <body>
-      <div class="panel heading">
-        <img src="/img/icon-48.png" width="48" height="48" alt="">
-        <h1>O que compra a sua hora - Customização</h1>
-      </div>
-      <div class="panel options-panel">
-        <label for="salario">Salário Mensal:</label>
-        <input class="salario" type="text" id="salario" name="salario" ><br><br>
-        <label for="horas-dia">Horas (por dia):</label>
-        <input class="horasDia" type="number" min="1" max="24" id="horas-dia" name="horas-dia"><br><br>
-        <label for="horas-semana">Horas (por semana):</label>
-        <input class="horasSemana" type="number" min="1" max="184" id="horas-semana" name="horas-semana"><br><br>
-        <div id="footer">
-          <button id="save-button">Salvar </button>
-          <span id="status"></span>
-        </div>
-      </div>
-    </body>
-    <script src="js/options.js"></script>
-  </html>
