--- manifest.json
@@ -1,34 +1,0 @@
-{
-   "update_url": "https://www.google.com",
-  
-   "name":"Mais-valia",
-   "short_name":"mais-valia",
-   "description":"Converte preços mostrados em quantas horas do seu tempo. Passar o mouse revela valor em reais",
-   "version":"0.1.0",
-   "icons":{  
-      "16":"img/icon-16.png",
-      "48":"img/icon-48.png",
-      "128":"img/icon-128.png"
-   },
-   "manifest_version":2,
-   "browser_action": {
-    "default_icon": "img/icon.png",
-    "default_popup": "index.html"
-   },
-   "permissions": ["activeTab","tabs","storage"],
-   "options_page": "mais-valia.html",
-   "background": {
-    "scripts": ["background.js"]
-   },
-   "content_scripts":[  
-      {  
-         "matches":[  
-            "*://*/mais-valia.html"
-         ],
-         "css":[
-            "css/style.css"
-         ],
-         "run_at":"document_end"
-      }
-   ]
-}
