+++ unit-test/src/modulesTest.js
@@ -1,0 +1,15 @@
+
+import * as FunctionModules from "../../extension/src/modules.js"
+
+function main(){
+
+    let response = FunctionModules.getSufixOfCurrency("mil");
+
+    console.log(response)
+
+}
+
+console.log("=========>");
+
+main();
+
